requirements.txt
Flask>=2.3,<4.0
Flask-Login>=0.6
matplotlib>=3.7,<3.9
pandas>=2.0,<2.3
numpy>=1.24,<2.0
requests>=2.28
firebase-admin>=6.5
google-cloud-firestore>=2.14,<3.0
